requirements-frozen-20260216.txt
absl-py==2.3.1
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.16.5
altair==5.5.0
altgraph @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
Authlib==1.6.8
backports-datetime-fromisoformat==2.0.3
backports.asyncio.runner==1.2.0
bandit==1.8.6
bcrypt==5.0.0
beautifulsoup4==4.14.3
bitarray==3.8.0
black==24.10.0
blinker==1.9.0
boolean.py==5.0
build==1.4.0
CacheControl==0.14.3
cachetools==6.2.6
ccxt==4.5.38
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.1.8
cloudpickle==3.1.2
coincurve==21.0.0
contourpy==1.3.0
coverage==7.10.7
cryptography==43.0.3
curl_cffi==0.13.0
cycler==0.12.1
cyclonedx-python-lib==9.1.0
cytoolz==1.1.0
dateparser==1.2.2
defusedxml==0.7.1
dparse==0.6.4
ecdsa==0.19.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
exceptiongroup==1.3.1
Farama-Notifications==0.0.4
fastapi==0.128.8
filelock==3.19.1
fonttools==4.60.2
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gitdb==4.0.12
GitPython==3.1.46
grpcio==1.78.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==1.1.1
h11==0.16.0
h5py==3.14.0
hexbytes==1.3.1
hf-xet==1.2.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.10.0
kiwisolver==1.4.7
librt==0.8.0
license-expression==30.4.4
loguru==0.7.3
lxml==5.4.0
macholib @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Mako==1.3.10
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multitasking==0.0.12
mypy==1.19.1
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.16.0
networkx==3.2.1
nltk==3.9.2
numpy==1.26.4
optree==0.18.0
packageurl-python==0.17.6
packaging==25.0
pandas==2.3.3
parsimonious==0.10.0
passlib==1.7.4
pathspec==1.0.4
peewee==3.19.0
pillow==11.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.4.0
plotly==6.5.2
pluggy==1.6.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.5
psycopg2-binary==2.9.11
py-serializable==2.1.0
pyarrow==21.0.0
pyasn1==0.6.2
pycares==4.11.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-asyncio==0.26.0
pytest-cov==6.3.0
pytest-mock==3.15.1
python-binance==1.0.34
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyunormalize==17.0.0
PyYAML==6.0.3
redis==5.3.1
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rich==13.9.4
rlp==4.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.19.1
ruff==0.15.1
safetensors==0.7.0
safety==3.7.0
safety-schemas==0.0.16
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
stable_baselines3==2.7.1
starlette==0.49.3
stevedore==5.5.0
streamlit==1.50.0
sympy==1.14.0
TA-Lib==0.6.8
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
tomlkit==0.14.0
toolz==1.1.0
torch==2.8.0
tornado==6.5.4
tqdm==4.67.3
transformers==4.57.6
typer==0.23.1
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
web3==7.14.1
websockets==14.2
Werkzeug==3.1.5
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
